+++ Source/JavaScriptCore/runtime/Options.cpp
@@ -605,7 +605,7 @@ static void overrideDefaults()
     Options::maximumInliningRecursion() = 3;
 #endif
 
-#if USE(BMALLOC_MEMORY_FOOTPRINT_API)
+#if USE(MEMORY_FOOTPRINT_API)
     // On iOS and conditionally Linux, we control heap growth using process memory footprint. Therefore these values can be agressive.
     Options::smallHeapRAMFraction() = 0.8;
     Options::mediumHeapRAMFraction() = 0.9;
