+++ Source/WTF/wtf/PlatformUse.h
@@ -264,8 +264,8 @@
 #define USE_UICONTEXTMENU 1
 #endif
 
-#if PLATFORM(IOS_FAMILY) || (!USE(SYSTEM_MALLOC) && (OS(LINUX) && (PLATFORM(GTK) || PLATFORM(WPE))))
-#define USE_BMALLOC_MEMORY_FOOTPRINT_API 1
+#if PLATFORM(IOS_FAMILY) || (OS(LINUX) && (PLATFORM(GTK) || PLATFORM(WPE)))
+#define USE_MEMORY_FOOTPRINT_API 1
 #endif
 
 #if !defined(USE_PLATFORM_REGISTERS_WITH_PROFILE) && OS(DARWIN) && CPU(ARM64) && defined(__LP64__)
