+++ Source/WebKit/WebProcess/cocoa/WebProcessCocoa.mm
@@ -144,7 +144,7 @@
 #if PLATFORM(IOS_FAMILY)
 #import "AccessibilityUtilitiesSPI.h"
 #import "UIKitSPI.h"
-#import <bmalloc/MemoryStatusSPI.h>
+#import <wtf/spi/darwin/MemoryStatusSPI.h>
 #endif
 
 #if PLATFORM(IOS_FAMILY)
