+++ Source/bmalloc/PlatformMac.cmake
@@ -4,7 +4,3 @@ list(APPEND bmalloc_SOURCES
     bmalloc/IsoHeap.cpp
     bmalloc/ProcessCheck.mm
 )
-
-list(APPEND bmalloc_PUBLIC_HEADERS
-    bmalloc/darwin/MemoryStatusSPI.h
-)
